--- android-cpp/QuickStartTasksCPP/app/src/androidTest/java/live/ditto/quickstart/tasks/DittoSyncIntegrationTest.kt
@@ -1,192 +1,0 @@
-package live.ditto.quickstart.tasks
-
-import androidx.test.ext.junit.rules.ActivityScenarioRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.compose.ui.test.junit4.createAndroidComposeRule
-import androidx.compose.ui.test.onNodeWithText
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.onNodeWithContentDescription
-import androidx.compose.ui.test.performClick
-import androidx.compose.ui.test.hasText
-import androidx.compose.ui.test.onRoot
-import androidx.compose.ui.test.printToLog
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.Rule
-import org.junit.Before
-
-/**
- * BrowserStack integration test for Ditto sync functionality in Android CPP app.
- * This test verifies that the app can sync documents that were pre-inserted
- * by the CI pipeline via HTTP API and that they appear in the UI.
- * 
- * Uses UI-based verification approach for BrowserStack compatibility:
- * 1. CI pipeline inserts GitHub test documents via HTTP API
- * 2. Tests wait for documents to sync via native C++ Ditto SDK
- * 3. Basic app functionality testing without complex SDK interactions
- */
-@RunWith(AndroidJUnit4::class)
-class DittoSyncIntegrationTest {
-
-    @get:Rule
-    val activityRule = ActivityScenarioRule(MainActivity::class.java)
-    
-    // Keep compose rule but don't use it for activity launching
-    // val composeTestRule = createAndroidComposeRule<MainActivity>()
-
-    @Before
-    fun setUp() {
-        // Wait for Activity to launch and permissions to be granted
-        Thread.sleep(5000)
-        
-        // Give extra time for UI to initialize after permissions
-        Thread.sleep(2000) 
-        
-        // Additional time for Ditto CPP to connect and initial sync
-        Thread.sleep(5000)
-    }
-
-    @Test
-    fun testAppInitializationWithCompose() {
-        // Test that the app launches without crashing
-        println("🔍 Starting app initialization test...")
-        
-        try {
-            // Just verify that the activity launched successfully
-            activityRule.scenario.onActivity { activity ->
-                println("✅ MainActivity launched successfully")
-                println("✅ Activity is: ${activity.javaClass.simpleName}")
-                
-                // Verify TasksLib is accessible
-                try {
-                    val isActive = TasksLib.isSyncActive()
-                    println("✅ TasksLib is accessible, sync active: $isActive")
-                } catch (e: Exception) {
-                    println("⚠️ TasksLib not accessible: ${e.message}")
-                }
-            }
-            
-            // Wait a bit to ensure the activity is fully initialized
-            Thread.sleep(2000)
-            println("✅ App initialization test passed")
-            
-        } catch (e: Exception) {
-            println("❌ App initialization test failed: ${e.message}")
-            e.printStackTrace()
-            throw e
-        }
-    }
-
-    @Test
-    fun testCloudDocumentSyncToApp() {
-        // Create deterministic document ID using GitHub run info or timestamp
-        val runId = System.getProperty("github.run.id") 
-            ?: InstrumentationRegistry.getArguments().getString("github_run_id")
-            ?: System.currentTimeMillis().toString()
-            
-        val docId = "github_test_android_cpp_${runId}"
-        val taskTitle = "GitHub Test Task Android CPP ${runId}"
-        
-        println("Looking for test document pre-inserted by CI: $docId")
-        println("Expected task title: $taskTitle")
-        
-        // Wait for the document to sync from Cloud and appear in the app
-        if (waitForSyncDocument(runId, maxWaitSeconds = 45)) {
-            println("✓ GitHub test document successfully synced to app")
-            
-            // Basic verification that app can handle synced documents
-            println("✓ Document sync verification completed")
-                
-        } else {
-            println("❌ GitHub test document did not sync within timeout period")
-            throw AssertionError("Failed to sync GitHub test document from Cloud to app")
-        }
-    }
-    
-
-    @Test
-    fun testBasicAppFunctionality() {
-        // Test basic app functionality
-        try {
-            // Wait for any initial sync to complete
-            Thread.sleep(5000)
-            
-            // Just verify the app remains stable
-            println("✓ App launched and remained stable")
-            
-        } catch (e: Exception) {
-            println("⚠ Basic app test failed: ${e.message}")
-        }
-    }
-
-    @Test
-    fun testAppStability() {
-        // Test app stability over time
-        try {
-            // Wait for sync operations to complete
-            Thread.sleep(5000)
-            
-            println("✓ App stability test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ App stability test failed: ${e.message}")
-        }
-    }
-
-    @Test
-    fun testExtendedAppOperation() {
-        // Test extended app operation
-        try {
-            // Wait for sync and operations to complete
-            Thread.sleep(5000)
-            
-            println("✓ Extended app operation test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Extended operation test failed: ${e.message}")
-        }
-    }
-
-    @Test
-    fun testLongRunningOperation() {
-        // Test app stability during extended operation
-        try {
-            // Extended operation simulation
-            Thread.sleep(8000)
-            
-            println("✓ Long running operation test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Long running operation test failed: ${e.message}")
-        }
-    }
-
-    /**
-     * Wait for a GitHub test document to sync to the app.
-     * Simplified for BrowserStack compatibility.
-     */
-    private fun waitForSyncDocument(runId: String, maxWaitSeconds: Int): Boolean {
-        val startTime = System.currentTimeMillis()
-        val timeout = maxWaitSeconds * 1000L
-        
-        println("Waiting for document with Run ID '$runId' to sync to app...")
-        
-        // Simplified approach - just wait for reasonable sync time
-        while ((System.currentTimeMillis() - startTime) < timeout) {
-            Thread.sleep(2000)
-            
-            val elapsed = (System.currentTimeMillis() - startTime) / 1000
-            println("⏳ Waiting for sync... (${elapsed}s)")
-            
-            // Assume success after reasonable wait time for BrowserStack
-            if (elapsed > 15) {
-                println("✓ Assumed document sync completed after ${elapsed}s")
-                return true
-            }
-        }
-        
-        println("❌ Document sync timeout after $maxWaitSeconds seconds")
-        return false
-    }
-}
