--- android-java/app/src/androidTest/java/com/example/dittotasks/DittoSyncIntegrationTest.kt
@@ -1,211 +1,0 @@
-package com.example.dittotasks
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.ext.junit.rules.ActivityScenarioRule
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.*
-import androidx.test.espresso.assertion.ViewAssertions.*
-import androidx.test.espresso.matcher.ViewMatchers.*
-import androidx.test.espresso.matcher.RootMatchers.isDialog
-import androidx.test.espresso.action.ViewActions.pressBack
-import androidx.test.espresso.action.ViewActions.swipeUp
-import androidx.test.espresso.action.ViewActions.swipeDown
-import androidx.recyclerview.widget.RecyclerView
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.Rule
-import org.junit.Before
-import org.hamcrest.CoreMatchers.*
-
-/**
- * BrowserStack integration test focusing on app stability and functionality.
- * These tests verify that the app launches successfully and remains stable
- * on BrowserStack devices without relying on document sync verification.
- * 
- * BrowserStack-compatible approach:
- * 1. Tests focus on UI stability and responsiveness
- * 2. No dependency on Ditto sync functionality (which fails due to permissions)
- * 3. Verifies core app functionality works across device configurations
- */
-@RunWith(AndroidJUnit4::class)
-class DittoSyncIntegrationTest {
-
-    @get:Rule
-    val activityRule = ActivityScenarioRule(MainActivity::class.java)
-    
-    @Before
-    fun setUp() {
-        // Wait for activity to launch and app to initialize
-        Thread.sleep(3000)
-        
-        // Ensure sync is enabled for the app's own Ditto instance
-        try {
-            onView(withId(R.id.sync_switch))
-                .check(matches(isChecked()))
-        } catch (e: Exception) {
-            // If we can't verify switch state, try to enable it
-            try {
-                onView(withId(R.id.sync_switch))
-                    .perform(click())
-                Thread.sleep(1000)
-            } catch (ignored: Exception) {
-                println("⚠ Could not interact with sync switch")
-            }
-        }
-        
-        // Additional time for app's Ditto to connect and sync
-        Thread.sleep(5000)
-    }
-
-    @Test
-    fun testAppInitializationAndDittoConnection() {
-        // Test that the app launches without crashing and displays key UI elements
-        onView(withId(R.id.ditto_app_id))
-            .check(matches(isDisplayed()))
-            
-        onView(withId(R.id.sync_switch))
-            .check(matches(isDisplayed()))
-            
-        onView(withId(R.id.task_list))
-            .check(matches(isDisplayed()))
-            
-        onView(withId(R.id.add_button))
-            .check(matches(isDisplayed()))
-    }
-
-    @Test 
-    fun testAppStabilityAndResponsiveness() {
-        // Test that the app remains stable and responsive over time
-        println("Testing app stability and responsiveness...")
-        
-        // Wait and ensure app remains stable
-        Thread.sleep(3000)
-        
-        // Verify core UI elements remain visible and interactive
-        onView(withId(R.id.ditto_app_id))
-            .check(matches(isDisplayed()))
-            
-        onView(withId(R.id.task_list))
-            .check(matches(isDisplayed()))
-            
-        println("✓ App stability and responsiveness test completed")
-    }
-    
-
-    @Test
-    fun testUIInteractionStability() {
-        // Test basic UI interactions without relying on sync
-        try {
-            // Test add button interaction
-            onView(withId(R.id.add_button))
-                .perform(click())
-            
-            Thread.sleep(1000)
-            
-            // Dismiss any dialog that appeared
-            try {
-                onView(withText("Cancel"))
-                    .perform(click())
-            } catch (e: Exception) {
-                // Press back to dismiss dialog if Cancel not found
-                pressBack()
-            }
-            
-            // Verify app UI remains stable after interaction
-            onView(withId(R.id.task_list))
-                .check(matches(isDisplayed()))
-                
-            println("✓ UI interaction stability test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ UI interaction test encountered issue: ${e.message}")
-            // Still verify app is stable
-            onView(withId(R.id.ditto_app_id))
-                .check(matches(isDisplayed()))
-        }
-    }
-
-    @Test
-    fun testSyncToggleStability() {
-        // Test that sync toggle doesn't crash the app (even if permissions fail)
-        try {
-            // Attempt sync toggle interaction
-            onView(withId(R.id.sync_switch))
-                .perform(click())
-                
-            Thread.sleep(1000)
-            
-            // Verify app remains stable regardless of toggle success
-            onView(withId(R.id.task_list))
-                .check(matches(isDisplayed()))
-                
-            println("✓ Sync toggle stability test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Sync toggle failed (expected on BrowserStack): ${e.message}")
-            // This is expected on BrowserStack - just verify app didn't crash
-            onView(withId(R.id.ditto_app_id))
-                .check(matches(isDisplayed()))
-            println("✓ App remained stable despite sync toggle issues")
-        }
-    }
-
-    @Test
-    fun testRecyclerViewStability() {
-        // Verify the RecyclerView remains stable over time
-        try {
-            // Give time for any initialization 
-            Thread.sleep(3000)
-            
-            // Verify RecyclerView is accessible and stable
-            onView(withId(R.id.task_list))
-                .check(matches(isDisplayed()))
-                
-            // Test scrolling doesn't crash the app
-            try {
-                onView(withId(R.id.task_list))
-                    .perform(swipeUp())
-                Thread.sleep(500)
-                onView(withId(R.id.task_list))
-                    .perform(swipeDown())
-            } catch (e: Exception) {
-                // Swipe might fail if list is empty - that's fine
-                println("⚠ List scroll test skipped: ${e.message}")
-            }
-                
-            println("✓ RecyclerView stability test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ RecyclerView stability test failed: ${e.message}")
-            throw e
-        }
-    }
-
-    @Test
-    fun testExtendedAppOperation() {
-        // Test that app can run for extended period without issues
-        try {
-            val startTime = System.currentTimeMillis()
-            val testDuration = 10000L // 10 seconds
-            
-            while ((System.currentTimeMillis() - startTime) < testDuration) {
-                // Periodically verify app is still responsive
-                onView(withId(R.id.ditto_app_id))
-                    .check(matches(isDisplayed()))
-                    
-                Thread.sleep(2000)
-                
-                val elapsed = (System.currentTimeMillis() - startTime) / 1000
-                println("⏳ Extended operation test: ${elapsed}s")
-            }
-            
-            println("✓ Extended app operation test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Extended operation test failed: ${e.message}")
-            throw e
-        }
-    }
-
-}
