--- android-kotlin/QuickStartTasks/app/src/androidTest/java/live/ditto/quickstart/tasks/DittoSyncIntegrationTest.kt
@@ -1,189 +1,0 @@
-package live.ditto.quickstart.tasks
-
-import androidx.test.ext.junit.rules.ActivityScenarioRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.compose.ui.test.junit4.createAndroidComposeRule
-import androidx.compose.ui.test.onNodeWithText
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.onNodeWithContentDescription
-import androidx.compose.ui.test.performClick
-import androidx.compose.ui.test.hasText
-import androidx.compose.ui.test.onRoot
-import androidx.compose.ui.test.printToLog
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.Rule
-import org.junit.Before
-
-/**
- * BrowserStack integration test focusing on app stability and functionality.
- * These tests verify that the app launches successfully and remains stable
- * on BrowserStack devices without relying on document sync verification.
- * 
- * BrowserStack-compatible approach:
- * 1. Tests focus on Compose UI stability and responsiveness
- * 2. No dependency on Ditto sync functionality (which fails due to permissions)
- * 3. Verifies core app functionality works across device configurations
- */
-@RunWith(AndroidJUnit4::class)
-class DittoSyncIntegrationTest {
-
-    @get:Rule
-    val composeTestRule = createAndroidComposeRule<MainActivity>()
-    
-    @Before
-    fun setUp() {
-        // Wait for Activity to launch and UI to initialize
-        Thread.sleep(3000)
-        
-        // Additional time for app's Ditto to connect and initial sync
-        Thread.sleep(5000)
-    }
-
-    @Test
-    fun testAppInitializationWithCompose() {
-        // Test that the app launches without crashing and displays key UI elements
-        composeTestRule.onNodeWithText("Tasks")
-            .assertIsDisplayed()
-            
-        composeTestRule.onNodeWithContentDescription("Add task")
-            .assertIsDisplayed()
-    }
-
-    @Test
-    fun testComposeUIStabilityAndResponsiveness() {
-        // Test that the Compose UI remains stable and responsive
-        println("Testing Compose UI stability and responsiveness...")
-        
-        // Wait for UI to stabilize
-        Thread.sleep(3000)
-        
-        // Print compose tree for debugging
-        composeTestRule.onRoot().printToLog("ComposeStabilityTest")
-        
-        // Verify core UI elements remain visible and functional
-        composeTestRule.onNodeWithText("Tasks")
-            .assertIsDisplayed()
-            
-        composeTestRule.onNodeWithContentDescription("Add task")
-            .assertIsDisplayed()
-            
-        println("✓ Compose UI stability and responsiveness test completed")
-    }
-    
-
-    @Test
-    fun testComposeUIInteractionStability() {
-        // Test basic Compose UI interactions without relying on sync
-        try {
-            // Test add button interaction
-            composeTestRule.onNodeWithContentDescription("Add task")
-                .performClick()
-            
-            Thread.sleep(1000)
-            
-            // Verify app UI remains stable after interaction
-            composeTestRule.onNodeWithText("Tasks")
-                .assertIsDisplayed()
-                
-            println("✓ Compose UI interaction stability test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Compose UI interaction test encountered issue: ${e.message}")
-            // Still verify core UI is stable
-            composeTestRule.onRoot().printToLog("ComposeErrorDebug")
-            // Basic verification that app didn't crash
-            Thread.sleep(1000)
-            println("✓ App remained stable despite interaction issues")
-        }
-    }
-
-    @Test
-    fun testComposeLayoutStability() {
-        // Test Compose layout stability over time
-        try {
-            // Wait for UI to fully render
-            Thread.sleep(3000)
-            
-            // Verify main UI elements remain stable
-            composeTestRule.onNodeWithText("Tasks")
-                .assertIsDisplayed()
-                
-            composeTestRule.onNodeWithContentDescription("Add task")
-                .assertIsDisplayed()
-                
-            // Test if UI can handle recomposition triggers
-            Thread.sleep(2000)
-            
-            composeTestRule.onNodeWithText("Tasks")
-                .assertIsDisplayed()
-                
-            println("✓ Compose layout stability test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Compose layout stability test failed: ${e.message}")
-            throw e
-        }
-    }
-
-    @Test
-    fun testComposeAnimationStability() {
-        // Test that Compose animations don't cause crashes
-        try {
-            // Wait and trigger any potential animations
-            Thread.sleep(2000)
-            
-            // Print compose tree
-            composeTestRule.onRoot().printToLog("ComposeAnimationTest")
-            
-            // Test interaction that might trigger animations
-            try {
-                composeTestRule.onNodeWithContentDescription("Add task")
-                    .performClick()
-                Thread.sleep(500)
-            } catch (e: Exception) {
-                println("⚠ Animation interaction failed: ${e.message}")
-            }
-            
-            // Verify UI remains stable
-            composeTestRule.onNodeWithText("Tasks")
-                .assertIsDisplayed()
-                
-            println("✓ Compose animation stability test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Animation stability test failed: ${e.message}")
-        }
-    }
-
-    @Test
-    fun testExtendedComposeOperation() {
-        // Test that Compose UI can run for extended period without issues
-        try {
-            val startTime = System.currentTimeMillis()
-            val testDuration = 10000L // 10 seconds
-            
-            while ((System.currentTimeMillis() - startTime) < testDuration) {
-                // Periodically verify Compose UI is still responsive
-                composeTestRule.onNodeWithText("Tasks")
-                    .assertIsDisplayed()
-                    
-                Thread.sleep(2000)
-                
-                val elapsed = (System.currentTimeMillis() - startTime) / 1000
-                println("⏳ Extended Compose operation test: ${elapsed}s")
-            }
-            
-            // Final verification
-            composeTestRule.onRoot().printToLog("ExtendedTestComplete")
-            
-            println("✓ Extended Compose operation test completed")
-            
-        } catch (e: Exception) {
-            println("⚠ Extended Compose operation test failed: ${e.message}")
-            throw e
-        }
-    }
-
-}
